# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "会議室管理システムのAPI",
        "title": "Elmo Project API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/rooms": {
            "get": {
                "description": "全ての会議室の一覧を取得します",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rooms"
                ],
                "summary": "会議室一覧を取得",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Room"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "新しい会議室を作成します",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rooms"
                ],
                "summary": "会議室を作成",
                "parameters": [
                    {
                        "description": "会議室情報",
                        "name": "room",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Room"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Room"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/rooms/{id}": {
            "get": {
                "description": "指定されたIDの会議室の詳細情報を取得します",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rooms"
                ],
                "summary": "会議室詳細を取得",
                "parameters": [
                    {
                        "type": "string",
                        "description": "会議室ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Room"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/rooms/{id}/conclusion": {
            "post": {
                "description": "指定された会議室の結論を保存します",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rooms"
                ],
                "summary": "会議の結論を保存",
                "parameters": [
                    {
                        "type": "string",
                        "description": "会議室ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "結論",
                        "name": "conclusion",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ConclusionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/rooms/{id}/result": {
            "get": {
                "description": "指定された会議室の結果情報（「それな」集計、チャットログ等）を取得します",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rooms"
                ],
                "summary": "会議結果を取得",
                "parameters": [
                    {
                        "type": "string",
                        "description": "会議室ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.RoomResultResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/rooms/{id}/sorena": {
            "post": {
                "description": "指定された会議室でユーザーの「それな」カウントを更新します",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rooms"
                ],
                "summary": "「それな」カウントを処理",
                "parameters": [
                    {
                        "type": "string",
                        "description": "会議室ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "「それな」情報",
                        "name": "sorena",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SorenaRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/rooms/{id}/start": {
            "post": {
                "description": "指定された会議室を開始し、AIが初期質問を生成します",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rooms"
                ],
                "summary": "会議室を開始",
                "parameters": [
                    {
                        "type": "string",
                        "description": "会議室ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.StartRoomResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/rooms/{id}/status": {
            "put": {
                "description": "指定された会議室のステータスを更新します",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rooms"
                ],
                "summary": "会議室のステータスを更新",
                "parameters": [
                    {
                        "type": "string",
                        "description": "会議室ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "ステータス情報",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateRoomStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/rooms/{id}/summary": {
            "post": {
                "description": "AIを使用して会議ログの要約を作成し、データベースに保存します",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rooms"
                ],
                "summary": "会議ログの要約を作成",
                "parameters": [
                    {
                        "type": "string",
                        "description": "会議室ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "要約対象のログ",
                        "name": "summary",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SummaryRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.ChatLog": {
            "type": "object",
            "properties": {
                "is_summary": {
                    "type": "boolean",
                    "example": false
                },
                "log_id": {
                    "type": "integer",
                    "example": 1
                },
                "message": {
                    "type": "string",
                    "example": "良いアイデアですね"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2024-01-01T10:00:00Z"
                },
                "user_id": {
                    "type": "string",
                    "example": "user123"
                }
            }
        },
        "models.ConclusionRequest": {
            "type": "object",
            "properties": {
                "conclusion": {
                    "type": "string",
                    "example": "来週までにプロトタイプを完成させる"
                }
            }
        },
        "models.LogEntry": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string",
                    "example": "プロジェクトの進捗について話し合いました"
                }
            }
        },
...
        "models.ParticipantUser": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "user123"
                },
                "name": {
                    "type": "string",
                    "example": "田中太郎"
                }
            }
        },
...
        "models.ResultRoomInfo": {
            "type": "object",
            "properties": {
                "room_id": {
                    "type": "string",
                    "example": "room123"
                },
                "title": {
                    "type": "string",
                    "example": "週次ミーティング"
                }
            }
        },
        "models.Room": {
            "type": "object",
            "properties": {
                "conclusion": {
                    "type": "string",
                    "example": "来週までにプロトタイプを完成させる"
                },
                "description": {
                    "type": "string",
                    "example": "今週の進捗確認と来週の計画"
                },
                "id": {
                    "type": "string",
                    "example": "abc123"
                },
                "initial_question": {
                    "type": "string",
                    "example": "今日の議題について何か質問はありますか？"
                },
                "status": {
                    "type": "string",
                    "example": "inprogress"
                },
                "title": {
                    "type": "string",
                    "example": "週次ミーティング"
                }
            }
        },
        "models.RoomInfo": {
            "type": "object",
            "properties": {
                "room_id": {
                    "type": "string",
                    "example": "room123"
                },
                "status": {
                    "type": "string",
                    "example": "inprogress"
                },
                "title": {
                    "type": "string",
                    "example": "週次ミーティング"
                }
            }
        },
        "models.RoomResultResponse": {
            "type": "object",
            "properties": {
                "chat_logs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ChatLog"
                    }
                },
                "room_info": {
                    "$ref": "#/definitions/models.ResultRoomInfo"
                },
                "sorena_summary": {
                    "$ref": "#/definitions/models.SorenaSummary"
                }
            }
        },
...
        "models.SorenaRequest": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer",
                    "example": 1
                },
                "user_id": {
                    "type": "string",
                    "example": "user123"
                }
            }
        },
        "models.SorenaSummary": {
            "type": "object",
            "properties": {
                "participants": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SorenaParticipant"
                    }
                },
                "total_count": {
                    "type": "integer",
                    "example": 15
                }
            }
        },
        "models.StartRoomResponse": {
            "type": "object",
            "properties": {
                "initial_question": {
                    "type": "string",
                    "example": "今日の議題について何か質問はありますか？"
                },
                "participants": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ParticipantUser"
                    }
                },
                "room_info": {
                    "$ref": "#/definitions/models.RoomInfo"
                }
            }
        },
        "models.SummaryRequest": {
            "type": "object",
            "properties": {
                "logs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.LogEntry"
                    }
                }
            }
        },
        "models.UpdateRoomStatusRequest": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "done"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}